+++ pom.xml
@@ -89,7 +89,7 @@
         <module>pubsub/cloud-client</module>
         <module>spanner/cloud-client</module>
         <module>speech/cloud-client</module>
-        <module>spring/integration/pubsub</module>
+        <module>spring/pubsub</module>
         <module>storage/cloud-client</module>
         <module>storage/json-api</module>
         <module>storage/storage-transfer</module>
