+++ spring/pubsub/pom.xml
@@ -6,7 +6,7 @@
     <artifactId>doc-samples</artifactId>
     <groupId>com.google.cloud</groupId>
     <version>1.0.0</version>
-    <relativePath>../../..</relativePath>
+    <relativePath>../..</relativePath>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <packaging>jar</packaging>
