+++ spring/pubsub/src/main/java/com/example/spring/pubsub/PubsubApplication.java
@@ -1,23 +1,29 @@
-package com.example.spring.pubsub;
+/*
+ *  Copyright 2017 original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 
-import java.io.IOException;
-import java.util.List;
-import java.util.stream.Collectors;
+package com.example.spring.pubsub;
 
 import com.google.cloud.pubsub.v1.AckReplyConsumer;
 import com.google.protobuf.ByteString;
-import com.google.pubsub.v1.Subscription;
-import com.google.pubsub.v1.SubscriptionName;
-import com.google.pubsub.v1.Topic;
-import com.google.pubsub.v1.TopicName;
+import java.io.IOException;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.gcp.pubsub.PubsubAdmin;
 import org.springframework.cloud.gcp.pubsub.core.PubsubTemplate;
 import org.springframework.cloud.gcp.pubsub.support.GcpHeaders;
 import org.springframework.cloud.gcp.pubsub.support.SubscriberFactory;
@@ -30,68 +36,39 @@ import org.springframework.integration.gcp.inbound.PubsubInboundChannelAdapter;
 import org.springframework.integration.gcp.outbound.PubsubMessageHandler;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.MessageHandler;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.view.RedirectView;
 
 @SpringBootApplication
-@RestController
 public class PubsubApplication {
 
   private static final Log LOGGER = LogFactory.getLog(PubsubApplication.class);
 
-  @Autowired
-  private PubsubOutboundGateway messagingGateway;
-
-  @Autowired
-  private PubsubAdmin admin;
-
   public static void main(String[] args) throws IOException {
     SpringApplication.run(PubsubApplication.class, args);
   }
-
-  @GetMapping("/listTopics")
-  public List<String> listTopics() {
-    return admin.listTopics().stream()
-        .map(Topic::getNameAsTopicName)
-        .map(TopicName::getTopic)
-        .collect(Collectors.toList());
-  }
-
-  @GetMapping("/listSubscriptions")
-  public List<String> listSubscriptions() {
-    return admin.listSubscriptions().stream()
-        .map(Subscription::getNameAsSubscriptionName)
-        .map(SubscriptionName::getSubscription)
-        .collect(Collectors.toList());
-  }
-
-  @PostMapping("/postMessage")
-  public RedirectView addMessage(@RequestParam("message") String message) {
-    messagingGateway.sendToPubsub(message);
-    return new RedirectView("/");
-  }
-
-  @PostMapping("/newTopic")
-  public RedirectView newTopic(@RequestParam("name") String topicName) {
-    admin.createTopic(topicName);
-    return new RedirectView("/");
-  }
-
-  @PostMapping("/newSubscription")
-  public RedirectView newSubscription(@RequestParam("name") String subscriptionName,
-      @RequestParam("topic") String topicName) {
-    admin.createSubscription(subscriptionName, topicName);
-    return new RedirectView("/");
-  }
-
+  /**
+   * Spring channel for incoming messages from Google Cloud Pub/Sub.
+   *
+   * <p>We use a {@link PublishSubscribeChannel} which broadcasts messages to every subscriber. In
+   * this case, every service activator.
+   */
   @Bean
   public MessageChannel pubsubInputChannel() {
     return new PublishSubscribeChannel();
   }
 
+  /**
+   * Inbound channel adapter that gets activated whenever a new message arrives at a Google Cloud
+   * Pub/Sub subscription.
+   *
+   * <p>Messages get posted to the specified input channel, which activates the service activators
+   * below.
+   *
+   * @param inputChannel Spring channel that receives messages and triggers attached service
+   * activators
+   * @param subscriberFactory creates the subscriber that listens to messages from Google Cloud
+   * Pub/Sub
+   * @return the inbound channel adapter for a Google Cloud Pub/Sub subscription
+   */
   @Bean
   public PubsubInboundChannelAdapter messageChannelAdapter(
       @Qualifier("pubsubInputChannel") MessageChannel inputChannel,
@@ -104,30 +81,46 @@ public class PubsubApplication {
     return adapter;
   }
 
+  /**
+   * Message handler that gets triggered whenever a new message arrives at the attached Spring
+   * channel.
+   *
+   * <p>Just logs the received message. Message acknowledgement mode set to manual above, so the
+   * consumer that allows us to (n)ack is extracted from the message headers and used to ack.
+   */
   @Bean
   @ServiceActivator(inputChannel = "pubsubInputChannel")
   public MessageHandler messageReceiver1() {
     return message -> {
       LOGGER.info("Message arrived! Payload: "
           + ((ByteString) message.getPayload()).toStringUtf8());
-      AckReplyConsumer consumer = (AckReplyConsumer) message.getHeaders().get(
-          GcpHeaders.ACKNOWLEDGEMENT);
+      AckReplyConsumer consumer =
+          (AckReplyConsumer) message.getHeaders().get(GcpHeaders.ACKNOWLEDGEMENT);
       consumer.ack();
     };
   }
 
+  /**
+   * Second message handler that also gets messages from the same subscription as above.
+   */
   @Bean
   @ServiceActivator(inputChannel = "pubsubInputChannel")
   public MessageHandler messageReceiver2() {
     return message -> {
       LOGGER.info("Message also arrived here! Payload: "
           + ((ByteString) message.getPayload()).toStringUtf8());
-      AckReplyConsumer consumer = (AckReplyConsumer) message.getHeaders().get(
-          GcpHeaders.ACKNOWLEDGEMENT);
+      AckReplyConsumer consumer =
+          (AckReplyConsumer) message.getHeaders().get(GcpHeaders.ACKNOWLEDGEMENT);
       consumer.ack();
     };
   }
 
+  /**
+   * The outbound channel adapter to write messages from a Spring channel to a Google Cloud Pub/Sub
+   * topic.
+   *
+   * @param pubsubTemplate Spring abstraction to send messages to Google Cloud Pub/Sub topics
+   */
   @Bean
   @ServiceActivator(inputChannel = "pubsubOutputChannel")
   public MessageHandler messageSender(PubsubTemplate pubsubTemplate) {
@@ -136,6 +129,9 @@ public class PubsubApplication {
     return outboundAdapter;
   }
 
+  /**
+   * A Spring mechanism to write messages to a channel.
+   */
   @MessagingGateway(defaultRequestChannel = "pubsubOutputChannel")
   public interface PubsubOutboundGateway {
