+++ spring/pubsub/src/main/wro/wro.properties
