+++ spring/pubsub/src/main/wro/wro.xml
